@@ -31,14 +31,14 @@
         "ext-fileinfo": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
-        "php": ">=7.4",
+        "php": "^8.0",
         "symfony/property-access": "^5.1",
         "symfony/serializer": "^5.1"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
         "php-coveralls/php-coveralls": "^2.2",
-        "phpunit/phpunit": "~9.3"
+        "phpunit/phpunit": "~9.4"
     },
     "scripts": {
         "lint-check": "php-cs-fixer fix --config=.php_cs.dist -v --diff --dry-run",
