@@ -56,14 +56,14 @@ class SerializerMediaTypeFormatterTest extends TestCase
     public function testReadingToStreamForUnsupportedTypeThrowsException(): void
     {
         $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage(\get_class($this->formatter) . ' cannot read type ' . self::class);
+        $this->expectExceptionMessage($this->formatter::class . ' cannot read type ' . self::class);
         $this->formatter->readFromStream(new Stream(fopen('php://temp', 'r+b')), self::class);
     }
 
     public function testWritingToStreamForUnsupportedTypeThrowsException(): void
     {
         $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage(\get_class($this->formatter) . ' cannot write type ' . self::class);
+        $this->expectExceptionMessage($this->formatter::class . ' cannot write type ' . self::class);
         $this->formatter->writeToStream($this, new Stream(fopen('php://temp', 'r+b')), null);
     }
 }
