@@ -19,15 +19,6 @@ use Aphiria\ContentNegotiation\MediaTypeFormatters\IMediaTypeFormatter;
  */
 final class ContentNegotiationResult
 {
-    /** @var IMediaTypeFormatter|null The matched media type formatter if there was one, otherwise null */
-    private ?IMediaTypeFormatter $formatter;
-    /** @var string|null The matched media type, or null if no media type was specified */
-    private ?string $mediaType;
-    /** @var string|null The matched encoding, or null if no encoding was specified */
-    private ?string $encoding;
-    /** @var string|null The matched language, or null if no language was matched */
-    private ?string $language;
-
     /**
      * @param IMediaTypeFormatter|null $formatter The matched media type formatter if there was one, otherwise null
      * @param string|null $mediaType The matched media type if there was one, otherwise null
@@ -35,15 +26,11 @@ final class ContentNegotiationResult
      * @param string|null $language The matched language if there was one, otherwise null
      */
     public function __construct(
-        ?IMediaTypeFormatter $formatter,
-        ?string $mediaType,
-        ?string $encoding,
-        ?string $language
+        private ?IMediaTypeFormatter $formatter,
+        private ?string $mediaType,
+        private ?string $encoding,
+        private ?string $language
     ) {
-        $this->formatter = $formatter;
-        $this->mediaType = $mediaType;
-        $this->encoding = $encoding;
-        $this->language = $language;
     }
 
     /**
