@@ -28,8 +28,6 @@ final class MediaTypeFormatterMatcher implements IMediaTypeFormatterMatcher
     private const FORMATTER_TYPE_INPUT = 'input';
     /** @const The type of formatter to match on for responses */
     private const FORMATTER_TYPE_OUTPUT = 'output';
-    /** @var IMediaTypeFormatter[] The list of supported media type formatters */
-    private array $mediaTypeFormatters;
     /** @var RequestHeaderParser The header parser to use to get request header values */
     private RequestHeaderParser $headerParser;
 
@@ -38,13 +36,12 @@ final class MediaTypeFormatterMatcher implements IMediaTypeFormatterMatcher
      * @param RequestHeaderParser|null $headerParser The header parser to use to get request header values
      * @throws InvalidArgumentException Thrown if there are no media type formatters specified
      */
-    public function __construct(array $mediaTypeFormatters, RequestHeaderParser $headerParser = null)
+    public function __construct(private array $mediaTypeFormatters, RequestHeaderParser $headerParser = null)
     {
-        if (\count($mediaTypeFormatters) === 0) {
+        if (\count($this->mediaTypeFormatters) === 0) {
             throw new InvalidArgumentException('List of formatters cannot be empty');
         }
 
-        $this->mediaTypeFormatters = $mediaTypeFormatters;
         $this->headerParser = $headerParser ?? new RequestHeaderParser();
     }
 
