@@ -33,8 +33,6 @@ class NegotiatedRequestBuilder extends RequestBuilder
 {
     /** @var IMediaTypeFormatterMatcher The media type formatter matcher */
     private IMediaTypeFormatterMatcher $mediaTypeFormatterMatcher;
-    /** @var string The default content type to use for bodies */
-    private string $defaultContentType;
 
     /**
      * @param IMediaTypeFormatterMatcher|null $mediaTypeFormatterMatcher The media type formatter matcher, or null if using the default one
@@ -43,7 +41,7 @@ class NegotiatedRequestBuilder extends RequestBuilder
      */
     public function __construct(
         IMediaTypeFormatterMatcher $mediaTypeFormatterMatcher = null,
-        string $defaultContentType = 'application/json',
+        private string $defaultContentType = 'application/json',
         string $defaultAccept = '*/*'
     ) {
         parent::__construct();
@@ -54,16 +52,15 @@ class NegotiatedRequestBuilder extends RequestBuilder
                 new HtmlMediaTypeFormatter(),
                 new PlainTextMediaTypeFormatter()
             ]);
-        $this->defaultContentType = $defaultContentType;
         $this->headers->add('Accept', $defaultAccept);
     }
 
     /**
      * @inheritdoc
-     * @param IBody|object|array|mixed $body The body to set
+     * @param mixed $body The body to set
      * @throws SerializationException Thrown if the body could not be serialized
      */
-    public function withBody($body): NegotiatedRequestBuilder
+    public function withBody(mixed $body): static
     {
         $new = clone $this;
 
