@@ -125,7 +125,7 @@ XML;
      * @param string $body The expected body of the stream
      * @return IStream|MockObject The stream that expects the input body
      */
-    private function createStreamThatExpectsBody(string $body): IStream
+    private function createStreamThatExpectsBody(string $body): IStream|MockObject
     {
         $stream = $this->createMock(IStream::class);
         $stream->expects($this->once())
@@ -141,7 +141,7 @@ XML;
      * @param string $body The body of the stream
      * @return IStream|MockObject The stream with the input body as its string body
      */
-    private function createStreamWithStringBody(string $body): IStream
+    private function createStreamWithStringBody(string $body): IStream|MockObject
     {
         $stream = $this->createMock(IStream::class);
         $stream->expects($this->once())
